+++ core/src/main/java/fr/uge/modules/linking/token/type/TypeDatetime.java
@@ -27,7 +27,7 @@ public class TypeDatetime implements TokenType{
     }
 
     @Override
-    public float computeProximity(List<TokenEntity> tokenLeft, List<TokenEntity> tokenRight) {
+    public double computeProximity(List<TokenEntity> tokenLeft, List<TokenEntity> tokenRight) {
         return 0;
     }
