+++ core/src/main/java/fr/uge/modules/linking/token/type/TypeEdgeResponse.java
@@ -1,6 +1,5 @@
 package fr.uge.modules.linking.token.type;
 
-import fr.uge.modules.api.model.TokenModel;
 import fr.uge.modules.api.model.entities.TokenEntity;
 
 import java.util.List;
@@ -27,7 +26,7 @@ public class TypeEdgeResponse implements TokenType {
     }
 
     @Override
-    public float computeProximity(List<TokenEntity> tokenLeft, List<TokenEntity> tokenRight) {
+    public double computeProximity(List<TokenEntity> tokenLeft, List<TokenEntity> tokenRight) {
         return 0;/*
         if (t1.equals(t2)) {
             return 100;
